Code cleanup changes to get rid of various compiler diagnostic messages.
git-svn-id: https://svn.apache.org/repos/asf/xerces/c/trunk@175709 13f79535-47bb-0310-9956-ffa450edef68

+++ src/xercesc/util/XMLDateTime.hpp
@@ -57,6 +57,9 @@
 /*
  * $Id$
  * $Log$
+ * Revision 1.17  2004/01/29 11:48:47  cargilld
+ * Code cleanup changes to get rid of various compiler diagnostic messages.
+ *
  * Revision 1.16  2004/01/13 19:50:56  peiyongz
  * remove parseContent()
  *
@@ -253,8 +256,8 @@ private:
     // -----------------------------------------------------------------------
     // Comparison
     // -----------------------------------------------------------------------
-    static int            compareResult(short
-                                      , short
+    static int            compareResult(int
+                                      , int
                                       , bool);
 
     static void           addDuration(XMLDateTime*             pDuration
